Claude optimized the display library

+++ dogm128_fast.h
@@ -54,7 +54,7 @@ void dogm128_line(int x0, int y0, int x1, int y1, dogm128_color_t color);
 void dogm128_rect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, dogm128_color_t color);
 void dogm128_fill_rect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, dogm128_color_t color);
 void dogm128_blit_aligned(uint8_t x, uint8_t y, const dogm128_bitmap_t *bmp);
-void dogm128_blit_or(uint8_t x, uint8_t y, const dogm128_bitmap_t *bmp);
+void dogm128_blit_or(int16_t x, int16_t y, const dogm128_bitmap_t *bmp);
 
 void dogm128_char(uint8_t x, uint8_t y, char c);
 void dogm128_text(uint8_t x, uint8_t y, const char *s);
